+++ command/server/config.go
@@ -60,8 +60,10 @@ func DevConfig(ha, transactional bool) *Config {
 			&Listener{
 				Type: "tcp",
 				Config: map[string]interface{}{
-					"address":     "127.0.0.1:8200",
-					"tls_disable": "1",
+					"address":                         "127.0.0.1:8200",
+					"tls_disable":                     true,
+					"proxy_protocol_behavior":         "allow_authorized",
+					"proxy_protocol_authorized_addrs": "127.0.0.1:8200",
 				},
 			},
 		},
@@ -666,6 +668,8 @@ func parseListeners(result *Config, list *ast.ObjectList) error {
 			"endpoint",
 			"infrastructure",
 			"node_id",
+			"proxy_protocol_behavior",
+			"proxy_protocol_authorized_addrs",
 			"tls_disable",
 			"tls_cert_file",
 			"tls_key_file",
@@ -673,6 +677,7 @@ func parseListeners(result *Config, list *ast.ObjectList) error {
 			"tls_cipher_suites",
 			"tls_prefer_server_cipher_suites",
 			"tls_require_and_verify_client_cert",
+			"tls_client_ca_file",
 			"token",
 		}
 		if err := checkHCLKeys(item.Val, valid); err != nil {