+++ command/server/listener.go
@@ -5,8 +5,10 @@ import (
 	// certificates that use it can be parsed.
 	_ "crypto/sha512"
 	"crypto/tls"
+	"crypto/x509"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"net"
 
 	"github.com/hashicorp/vault/helper/parseutil"
@@ -103,6 +105,18 @@ func listenerWrapTLS(
 		}
 		if requireClient {
 			tlsConf.ClientAuth = tls.RequireAndVerifyClientCert
+		}
+		if tlsClientCaFile, ok := config["tls_client_ca_file"]; ok {
+			caPool := x509.NewCertPool()
+			data, err := ioutil.ReadFile(tlsClientCaFile.(string))
+			if err != nil {
+				return nil, nil, nil, fmt.Errorf("failed to read tls_client_ca_file: %v", err)
+			}
+
+			if !caPool.AppendCertsFromPEM(data) {
+				return nil, nil, nil, fmt.Errorf("failed to parse CA certificate in tls_client_ca_file")
+			}
+			tlsConf.ClientCAs = caPool
 		}
 	}