+++ vault/barrier_aes_gcm_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/hashicorp/vault/helper/logformat"
 	"github.com/hashicorp/vault/physical"
+	"github.com/hashicorp/vault/physical/inmem"
 	log "github.com/mgutz/logxi/v1"
 )
 
@@ -16,8 +17,10 @@ var (
 
 // mockBarrier returns a physical backend, security barrier, and master key
 func mockBarrier(t testing.TB) (physical.Backend, SecurityBarrier, []byte) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -31,8 +34,10 @@ func mockBarrier(t testing.TB) (physical.Backend, SecurityBarrier, []byte) {
 }
 
 func TestAESGCMBarrier_Basic(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -41,8 +46,10 @@ func TestAESGCMBarrier_Basic(t *testing.T) {
 }
 
 func TestAESGCMBarrier_Rotate(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -51,8 +58,10 @@ func TestAESGCMBarrier_Rotate(t *testing.T) {
 }
 
 func TestAESGCMBarrier_Upgrade(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b1, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -65,8 +74,10 @@ func TestAESGCMBarrier_Upgrade(t *testing.T) {
 }
 
 func TestAESGCMBarrier_Upgrade_Rekey(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b1, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -79,8 +90,10 @@ func TestAESGCMBarrier_Upgrade_Rekey(t *testing.T) {
 }
 
 func TestAESGCMBarrier_Rekey(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -91,8 +104,10 @@ func TestAESGCMBarrier_Rekey(t *testing.T) {
 // Test an upgrade from the old (0.1) barrier/init to the new
 // core/keyring style
 func TestAESGCMBarrier_BackwardsCompatible(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -171,8 +186,10 @@ func TestAESGCMBarrier_BackwardsCompatible(t *testing.T) {
 
 // Verify data sent through is encrypted
 func TestAESGCMBarrier_Confidential(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -209,8 +226,10 @@ func TestAESGCMBarrier_Confidential(t *testing.T) {
 
 // Verify data sent through cannot be tampered with
 func TestAESGCMBarrier_Integrity(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -245,8 +264,10 @@ func TestAESGCMBarrier_Integrity(t *testing.T) {
 
 // Verify data sent through cannot be moved
 func TestAESGCMBarrier_MoveIntegrityV1(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -287,8 +308,10 @@ func TestAESGCMBarrier_MoveIntegrityV1(t *testing.T) {
 }
 
 func TestAESGCMBarrier_MoveIntegrityV2(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -329,8 +352,10 @@ func TestAESGCMBarrier_MoveIntegrityV2(t *testing.T) {
 }
 
 func TestAESGCMBarrier_UpgradeV1toV2(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -382,8 +407,10 @@ func TestAESGCMBarrier_UpgradeV1toV2(t *testing.T) {
 }
 
 func TestEncrypt_Unique(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -410,8 +437,10 @@ func TestEncrypt_Unique(t *testing.T) {
 }
 
 func TestInitialize_KeyLength(t *testing.T) {
-
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -441,7 +470,13 @@ func TestInitialize_KeyLength(t *testing.T) {
 }
 
 func TestEncrypt_BarrierEncryptor(t *testing.T) {
-	inm := physical.NewInmem(logger)
+	inm, err := inmem.NewInmem(nil, logger)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
 	b, err := NewAESGCMBarrier(inm)
 	if err != nil {
 		t.Fatalf("err: %v", err)